Changed the routes for the /profile/username/ and moved the html file to the main file. Also, incorporated Ebenezer's merged code about vounteer profile to our coe before we moved everything to the main file.
